requirements.txt
flask==3.1.0
yfinance==0.2.55
flask-socketio==5.3.0
apscheduler==3.10.4
alpha-vantage==2.3.1
finnhub-python==2.4.20
requests==2.31.0
python-dotenv==1.0.0